requirements.txt
selenium==4.16.0
pyinstaller==6.18.0
pyyaml==6.0.1
schedule==1.2.0
requests==2.31.0
PyGithub==2.1.1
packaging==24.0
psutil==5.9.6
